requirements.txt
pyserial==3.2.1
requests==2.12.4
beautifulsoup4==4.5.3
